backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
nltk==3.8.1 